requirements.txt
# dreamer-car — full dependency list
# See CLAUDE.md "Setup" section for the complete install sequence.
# Do NOT run this file alone — read CLAUDE.md first (PyTorch and two editable
# installs must be done separately before pip install -r requirements.txt).

# ── PyTorch (install FIRST with the right index-url for your hardware) ────────
# macOS MPS / CPU:    pip install torch torchvision
# Linux CUDA 11.8:    pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# Linux CUDA 12.1:    pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
torch>=2.1.0
torchvision>=0.16.0

# ── DreamerV3 ─────────────────────────────────────────────────────────────────
numpy>=1.24.0
Pillow>=9.0.0
einops>=0.6.0

# ── Logging ───────────────────────────────────────────────────────────────────
wandb>=0.15.0
tensorboard>=2.13.0
imageio>=2.28.0

# ── Donkeycar runtime deps (manage.py drive) ──────────────────────────────────
# donkeycar and gym-donkeycar themselves must be editable installs — see CLAUDE.md
docopt
tornado
paho-mqtt
simple-pid
progress
pyfiglet
psutil
pynmea2
pyserial
utm
pandas
pyyaml
prettytable
requests

# ── Gym (required by gym-donkeycar / dgym.py) ─────────────────────────────────
# Must be pinned — gym-donkeycar 1.3.1 imports `gym` directly (not gymnasium)
gym==0.22.0
gymnasium>=0.29.0
